Preparations Part 1
After some testing with tic-tac-toe, I have a successful model running in stable-baselines3 that should be able to attempt to learn any environment using PPO.
Importantly, this includes action masking, which was built off of David Foster's SIMPLE project. The legal actions in a state are sent as a boolean vector appended to the observation. The custom network takes this in and separates the two, using the legal actions as a mask of the logits output by the policy head.

The other major change is the creation of a game randomizer. In order to improve training quality, I created a method that resets and then randomly sets up a game of ROOT. Hopefully this will give the models a more diverse set of games to deal with and allow them to develop more robust strategies.

Other changes:
- Allowed the user to set a max number of actions when creating the environment
- Fixed a bug where the if in a battle, the attacker had a choice of what to damage when taking hits, they would end up destroying all of their cardboard in that clearing instead of the exact number of hits left to take (forgot a "- 1")
- Fixed a bug where if the Vagabond was coalitioned with the Woodland Alliance, they were not allowed to do the 'strike' action on any Eyrie targets