File System Complete
Altering the file system from SIMPLE, I now have a system in place to start training of the Marquise model from the command line. A lot of big changes were made, including a reworked wrapper around the given environment to manage the actions performed by the different agents, even those that are not learning at the time (just acting as part of the environment itself).

Main changes:
- Config file now contains action/obs sizes
- Training environment uses a get_main_observation method that is necessary to use for truncated games to return a fitting observation for the main model
- Made a simple "rules" algorithm that can play a faction by returning a distribution across legal actions given the current environment
- Changed rewards to be +1 and -1 for a win or loss
- Updated the main root environment to tell terminated vs truncated, set custom turn limits on creation
- Fixed an old bug that caused the VB to go hostile in the wrong order, causing them to miss out on points from infamy from cardboard
- Imported all of the SIMPLE files and reworked the ones needed for train.py to work