Beginning Training
There is now one environment "MainBaseEnv" that can be used by any of the factions to train on. Given who the "main player" to train is, the environment will only give that player's observation when it is their turn, saving time on the "rules" opponents who do not need an observation constructed.

The "rules" method has been updated to include the cats and uses the correct "skip" actions for each faction.

The test.py script has been updated to function for one faction model competing against rules opponents.