Mixing of Models
Simplified the setup under the hood, allowing more customization to be done from the train/test files.

The gymnasium environment that interacts with the RootGame object must now be told which rules opponent to use, along with which factions are being controlled by a model.

Models are now organized and named in a specific way. Each model has a type, detailed in 3 parts:
1. The faction name, capitalized and fully spelled out, like Marquise
2. The goal of the model. This is 'Main' for the default selfish model, or something like 'HelpsVB' for an alternate goal model.
3. An underscore, followed by the name of the rules type the model is to be trained in, such as 'Algo' or 'Starter'

The TrainingModelManager now looks at the model type to determine what type of model is being trained. For example, a Marquise model that only wants the Eyrie to lose the game. This will alter the goal

- The rules type given to the RootEnv object is used to control starting aspects of the training games, such as the number of turns allowed or args to be passed to the Game object
- The specific type of model + training type name used in the folder names is constructed from the CLI
- Forgot to add the loop that keeps track of rewards accumulated on opponent turns to return to the training model
- git ignoring the zoo for now as the number of models saved increases greatly